some more clean up
